+++ Tools/SVStatTool.cpp
@@ -366,7 +366,7 @@ DWORD SVStatisticsToolClass::FreeResult (SVStatisticsFeatureEnum aFeatureIndex)
 	{
 		SvStl::MessageMgrStd MesMan(SvStl::MsgType::Log );
 		MesMan.setMessage( SVMSG_SVO_103_REPLACE_ERROR_TRAP, SvStl::Tid_UnexpectedError, SvStl::SourceFileParams(StdMessageParams), SvStl::Err_16208);
-		LastError = -SvStl::Err_16208;
+		LastError = static_cast<DWORD> (-SvStl::Err_16208);
 	}
 		
 	return LastError;
